chore(extension): setup new Home component, LEA-3291
