+++ apps/extension/src/app/pages/home/home.tsx
@@ -1,104 +1,9 @@
-import { Route, useNavigate } from 'react-router';
+import { useFlags } from '@app/features/feature-flags';
 
-import { HomePageSelectors } from '@tests/selectors/home.selectors';
-import { Box, Stack } from 'leather-styles/jsx';
-
-import { RouteUrls } from '@shared/route-urls';
-
-import { useAccountScaledBalanceAnalytics } from '@app/common/app-analytics';
-import { formatCurrency } from '@app/common/currency-formatter';
-import { useAccountDisplayName } from '@app/common/hooks/account/use-account-names';
-import { useOnboardingState } from '@app/common/hooks/auth/use-onboarding-state';
-import { useOnMount } from '@app/common/hooks/use-on-mount';
-import { useSwitchAccountSheet } from '@app/common/switch-account/use-switch-account-sheet-context';
-import { whenPageMode } from '@app/common/utils';
-import { ActivityList } from '@app/features/activity-list/activity-list';
-import { FeedbackButton } from '@app/features/feedback-button/feedback-button';
-import { PromoBanner } from '@app/features/promo-banner/promo-banner';
-import { Assets } from '@app/pages/home/components/assets';
-import { useCurrentAccountTotalBalance } from '@app/query/common/account-balance/account-balance.query';
-import { homePageModalRoutes } from '@app/routes/app-routes';
-import { ModalBackgroundWrapper } from '@app/routes/components/modal-background-wrapper';
-import { useCurrentAccountIndex } from '@app/store/accounts/account';
-import { useCurrentStacksAccount } from '@app/store/accounts/blockchain/stacks/stacks-account.hooks';
-import {
-  refreshLeatherTabs,
-  useOnFinishedOnboarding,
-} from '@app/store/onboarding/onboarding.hooks';
-import { useTogglePrivateMode } from '@app/store/settings/settings.actions';
-import { useIsPrivateMode } from '@app/store/settings/settings.selectors';
-import { AccountCard } from '@app/ui/components/account/account.card';
-
-import { AccountActions } from './components/account-actions';
-import { HomeTabs } from './components/home-tabs';
-
-const emptyBalanceDisplay = '-.--';
+import { HomeV1 } from './home-v1';
+import { HomeV2 } from './home-v2';
 
 export function Home() {
-  const { decodedAuthRequest } = useOnboardingState();
-  const { toggleSwitchAccount } = useSwitchAccountSheet();
-  const navigate = useNavigate();
-  const account = useCurrentStacksAccount();
-  const currentAccountIndex = useCurrentAccountIndex();
-  const isPrivateMode = useIsPrivateMode();
-  const togglePrivateMode = useTogglePrivateMode();
-  useAccountScaledBalanceAnalytics({ accountIndex: currentAccountIndex });
-
-  useOnFinishedOnboarding(() => refreshLeatherTabs());
-
-  const { data: name = '', isFetching: isFetchingBnsName } = useAccountDisplayName({
-    address: account?.address || '',
-    index: currentAccountIndex || 0,
-  });
-
-  const balance = useCurrentAccountTotalBalance();
-
-  useOnMount(() => {
-    if (decodedAuthRequest) return navigate(RouteUrls.ChooseAccount);
-  });
-
-  return (
-    <Stack
-      data-testid={HomePageSelectors.HomePageContainer}
-      px={{ base: 0, md: 'space.05' }}
-      py={{ base: 0, md: 'space.07' }}
-      gap={{ base: 0, md: 'space.06' }}
-      width="100%"
-      bg="ink.1"
-      borderRadius="lg"
-      animation="fadein"
-      animationDuration="500ms"
-    >
-      <Box px={{ base: 'space.05', md: 0 }} pb={{ base: 'space.05', md: 0 }}>
-        <AccountCard
-          name={name}
-          availableBalance={
-            balance.state !== 'success' ? emptyBalanceDisplay : formatCurrency(balance.value)
-          }
-          totalBalance={
-            balance.state !== 'success' ? emptyBalanceDisplay : formatCurrency(balance.value)
-          }
-          toggleSwitchAccount={() => toggleSwitchAccount()}
-          isFetchingBnsName={isFetchingBnsName}
-          isLoadingBalance={balance.state === 'loading'}
-          isLoadingAdditionalData={balance.state === 'loading'}
-          isBalancePrivate={isPrivateMode}
-          onShowBalance={togglePrivateMode}
-        >
-          <AccountActions />
-        </AccountCard>
-        <PromoBanner />
-      </Box>
-      {whenPageMode({ full: <FeedbackButton />, popup: null })}
-      <HomeTabs>
-        <ModalBackgroundWrapper>
-          <Route index element={<Assets />} />
-          <Route path={RouteUrls.Activity} element={<ActivityList />}>
-            {homePageModalRoutes}
-          </Route>
-          {homePageModalRoutes}
-        </ModalBackgroundWrapper>
-      </HomeTabs>
-    </Stack>
-  );
+  const { extension_revamp } = useFlags();
+  return extension_revamp ? <HomeV2 /> : <HomeV1 />;
 }
